// src/ErrorDashboard/Client/src/api/types.gen.ts
export type DailyPointResponseModel = {
    date: string;
    count: number;
    weightedCount: number;
    estimatedRequests: number;
};